add modify figure
