+++ src/Form/Figure/FigureType.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Form\Figure;
+
+use App\Entity\Figure;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class FigureType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $builder
+            ->add('name', TextType::class)
+            ->add('description', TextareaType::class)
+            ->add('figure_group', TextType::class)
+            ->add('picture', TextType::class)
+            ->add('video', TextType::class)
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults([
+            'data_class' => Figure::class,
+        ]);
+    }
+}
