--- src/Form/Type/FigureType.php
@@ -1,57 +1,0 @@
-<?php
-
-
-namespace App\Form\Type;
-
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\DateType;
-use Symfony\Component\Form\Extension\Core\Type\SubmitType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\FormBuilderInterface;
-
-class FigureType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('name', TextType::class, [
-                'attr' => [
-                    'placeholder' => 'Nom',
-                    'class' =>'appearance-none block w-full bg-white text-gray-700 border border-gray-400 shadow-inner rounded-md py-3 px-4 leading-tight focus:outline-none  focus:border-gray-500'
-                ]
-            ])
-            ->add('description', TextareaType::class, [
-                'attr' => [
-                    'placeholder' => 'Description',
-                    'class' => 'appearance-none block w-full bg-white text-gray-700 border border-gray-400 shadow-inner rounded-md py-3 px-4 leading-tight focus:outline-none  focus:border-gray-500'
-                ]
-            ])
-            ->add('figure_group', TextType::class, [
-                'attr' => [
-                    'placeholder' => 'Figure group',
-                    'class' => 'block appearance-none text-gray-600 w-full bg-white border border-gray-400 shadow-inner px-4 py-2 pr-8 rounded',
-                ]
-            ])
-            ->add('picture', TextType::class,
-                [
-                    'attr' => [
-                        'placeholder' => 'Photos',
-                        'class' => 'appearance-none block w-full bg-white text-gray-700 border border-gray-400 shadow-inner rounded-md py-3 px-4 leading-tight focus:outline-none  focus:border-gray-500'
-                    ]
-                ])
-            ->add('video', TextType::class,
-                [
-                    'attr' => [
-                        'placeholder' => 'Videos',
-                        'class' => 'appearance-none block w-full bg-white text-gray-700 border border-gray-400 shadow-inner rounded-md py-3 px-4 leading-tight focus:outline-none  focus:border-gray-500'
-                    ]
-                ])
-            ->add('save', SubmitType::class, [
-                'label' => 'Enregistrer',
-                'attr' => [
-                    'class' => 'appearance-none bg-gray-200 text-gray-900 px-2 py-1 shadow-sm border border-gray-400 rounded-md mr-3'
-                ]
-            ]);
-    }
-}
