+++ templates/figure/formFigure.html.twig
@@ -14,37 +14,46 @@
                         <label class='block uppercase tracking-wide text-xs font-bold mb-2'
                                for='nom'>Nom</label>
                         <div id='nom'>
-                            {{ form_widget(formFigure.name) }}
+                            {{ form_widget(formFigure.name, {'attr': {'class' : 'appearance-none block w-full bg-white text-gray-700 border border-gray-400 shadow-inner rounded-md py-3 px-4 leading-tight focus:outline-none  focus:border-gray-500'}}) }}
                         </div>
                     </div>
                     <div class='w-full md:w-full px-3 mb-6'>
                         <label class='block uppercase tracking-wide text-xs font-bold mb-2'
                                for='description'>Description</label>
                         <div id='description'>
-                            {{ form_widget(formFigure.description) }}
+                            {{ form_widget(formFigure.description, {'attr': {'class' : 'appearance-none block w-full bg-white text-gray-700 border border-gray-400 shadow-inner rounded-md py-3 px-4 leading-tight focus:outline-none  focus:border-gray-500'}}) }}
                         </div>
                     </div>
                     <div class='w-full md:w-full px-3 mb-6'>
                         <label class='block uppercase tracking-wide text-xs font-bold mb-2'
                                for='figure_group'>Figure Group</label>
                         <div id='figure_group'>
-                            {{ form_widget(formFigure.figure_group) }}
+                            {{ form_widget(formFigure.figure_group, {'attr': {'class' : 'appearance-none block w-full bg-white text-gray-700 border border-gray-400 shadow-inner rounded-md py-3 px-4 leading-tight focus:outline-none  focus:border-gray-500'}}) }}
                         </div>
                     </div>
                     <div class='w-full md:w-full px-3 mb-6'>
                         <label class='block uppercase tracking-wide text-xs font-bold mb-2'
                                for='photo'>Photos</label>
-                        <div id='photo'>{{ form_widget(formFigure.picture) }}
+                        <div id='photo'>{{ form_widget(formFigure.picture, {'attr': {'class' : 'appearance-none block w-full bg-white text-gray-700 border border-gray-400 shadow-inner rounded-md py-3 px-4 leading-tight focus:outline-none  focus:border-gray-500'}}) }}
                         </div>
                     </div>
                     <div class='w-full md:w-full px-3 mb-6'>
                         <label class='block uppercase tracking-wide text-xs font-bold mb-2'
                                for='video'>Videos</label>
                         <div id='video'>
-                            {{ form_widget(formFigure.video) }}
+                            {{ form_widget(formFigure.video, {'attr': {'class' : 'appearance-none block w-full bg-white text-gray-700 border border-gray-400 shadow-inner rounded-md py-3 px-4 leading-tight focus:outline-none  focus:border-gray-500'}}) }}
                         </div>
                     </div>
-                    {{ form_widget(formFigure.save) }}
+                    <div class="flex justify-end">
+                        <button class="appearance-none bg-gray-200 text-gray-900 px-2 py-1 shadow-sm border border-gray-400 rounded-md mr-3"
+                                type="submit">
+                            {% if editMode %}
+                                Modifier
+                            {% else %}
+                                Enregistrer
+                            {% endif %}
+                        </button>
+                    </div>
                     {{ form_end(formFigure) }}
                 </div>
             </div>
